requirement.txt
absl-py==0.6.1
alembic==1.0.0
amqp==1.4.9
aniso8601==4.0.1
anyjson==0.3.3
asn1crypto==0.24.0
astor==0.7.1
async==0.6.2
asyncio==3.4.3
attrs==18.2.0
Automat==0.7.0
backcall==0.1.0
beautifulsoup4==4.6.3
billiard==3.3.0.23
bleach==3.0.2
blinker==1.4
boolexp==0.1
bs4==0.0.1
celery==3.1.26.post2
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
click==6.7
colorama==0.3.9
colorlog==3.1.4
constantly==15.1.0
cryptography==2.3.1
cssselect==1.0.3
decorator==4.3.0
defusedxml==0.5.0
distro==1.3.0
Django==1.11.7
django-celery==3.2.2
django-redis==4.10.0
django-tinymce==2.7.0
djangorestframework==3.9.0
dominate==2.3.4
entrypoints==0.2.3
Flask==1.0.2
flask-blueprint==1.2.2
Flask-Bootstrap==3.3.7.1
Flask-Cache==0.13.1
Flask-Caching==1.4.0
Flask-HTTPAuth==3.2.4
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.2.1
Flask-Moment==0.6.0
Flask-RESTful==0.3.6
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
Flask-Uploads==0.2.1
Flask-WTF==0.14.2
future==0.17.1
gast==0.2.0
grpcio==1.16.1
hyperlink==18.0.0
idna==2.7
incremental==17.5.0
ipykernel==5.1.0
ipython==7.1.1
ipython-genutils==0.2.0
ipywidgets==7.4.2
itsdangerous==0.24
jedi==0.13.1
Jinja2==2.10
jsonpath==0.75
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==6.0.0
jupyter-core==4.4.0
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
kombu==3.0.37
lxml==4.2.5
Mako==1.0.7
Markdown==3.0.1
MarkupSafe==1.0
mistune==0.8.4
nbconvert==5.4.0
nbformat==4.4.0
networkx==2.2
notebook==5.7.0
numpy==1.15.4
pandocfilters==1.4.2
parsel==1.5.0
parso==0.3.1
pickleshare==0.7.5
Pillow==5.3.0
prometheus-client==0.4.2
prompt-toolkit==2.0.7
protobuf==3.6.1
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycparser==2.18
PyDispatcher==2.0.5
pygame==1.9.4
Pygments==2.2.0
PyHamcrest==1.9.0
PyMySQL==0.9.2
pyOpenSSL==18.0.0
pyparsing==2.2.1
python-dateutil==2.7.3
python-editor==1.0.3
pytz==2018.5
pywinpty==0.5.4
pyzmq==17.1.2
qtconsole==4.4.2
queuelib==1.5.0
redis==2.10.6
requests==2.19.1
ruamel.yaml==0.15.70
Scrapy==1.5.1
selenium==3.14.1
Send2Trash==1.5.0
serializers==0.2.4
service-identity==17.0.0
six==1.11.0
SQLAlchemy==1.2.11
tabulate==0.8.2
tensorboard==1.12.0
tensorflow==1.12.0
termcolor==1.1.0
terminado==0.8.1
testpath==0.4.2
tinycm==0.1.29
tornado==5.1.1
traitlets==4.3.2
Twisted==17.9.0
urllib3==1.23
vine==1.1.4
visitor==0.1.3
w3lib==1.19.0
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.4.2
WTForms==2.2.1
zope.interface==4.5.0
